refactor: modularize build job execution logic and remove custom workflow script fields from database schema
