+++ apps/openci_build_job_processor/lib/openci_build_job_processor.dart
@@ -1,9 +1,12 @@
 library;
 
+export 'src/build_job/build_job_executor.dart';
+export 'src/build_job/cleanup_build_job_workspace.dart';
+export 'src/build_job/prepare_build_job_workspace.dart';
+export 'src/build_job/run_build_job_script.dart';
 export 'src/build_job_poller/job_poller.dart';
-export 'src/job_executor.dart';
-export 'src/processor_config.dart';
-export 'src/sentry.dart';
 export 'src/logging/build_step_logger.dart';
 export 'src/orchard/orchard_api_client.dart';
 export 'src/orchard/orchard_vm_service.dart';
+export 'src/processor_config.dart';
+export 'src/sentry.dart';
