+++ apps/openci_build_job_processor/lib/src/build_job_poller/claim_next_job.dart
@@ -4,12 +4,10 @@ Future<BuildJob?> claimNextJob({
   required OpenCiApiService apiService,
   String? vmName,
   String? workerHost,
-  int? maxConcurrentJobs,
 }) async {
   final request = ClaimJobRequest(
     vmName: vmName,
     workerHost: workerHost,
-    maxConcurrentJobs: maxConcurrentJobs,
   );
 
   final response = await apiService.claimNextJob(request.toJson());
