--- apps/openci_build_job_processor/lib/src/build_job_poller/wait_for_available_slot.dart
@@ -1,15 +1,0 @@
-import 'package:logging/logging.dart';
-
-Future<void> waitForAvailableSlot({
-  Duration delay = const Duration(seconds: 5),
-  required int Function() getActiveJobsCount,
-  required int maxConcurrentJobs,
-  required Logger log,
-}) async {
-  while (getActiveJobsCount() >= maxConcurrentJobs) {
-    log.info(
-      'Active jobs ${getActiveJobsCount()} >= max $maxConcurrentJobs. Waiting...',
-    );
-    await Future<void>.delayed(delay);
-  }
-}
