+++ apps/openci_server/lib/build_job/build_job.dart
@@ -54,8 +54,6 @@ class BuildJobs extends Table {
   TextColumn get workflowRunId => text().nullable()();
   TextColumn get needs => text().map(const StringListConverter()).nullable()();
   TextColumn get runsOn => text().nullable()();
-  TextColumn get customScript => text().nullable()();
-  TextColumn get workflowYaml => text().nullable()();
 
   TextColumn get failureSummary => text().nullable()();
   TextColumn get failureSummaryModel => text().nullable()();
