+++ apps/openci_server/lib/database.dart
@@ -64,13 +64,7 @@ class AppDatabase extends _$AppDatabase {
   @override
   MigrationStrategy get migration {
     return MigrationStrategy(
-      beforeOpen: (details) async {
-        try {
-          await customStatement(
-            'ALTER TABLE build_jobs ADD COLUMN IF NOT EXISTS workflow_yaml TEXT;',
-          );
-        } catch (_) {}
-      },
+      beforeOpen: (details) async {},
       onCreate: (Migrator m) async {
         await m.createAll();
       },
