+++ apps/openci_server/lib/seed/seed_dao.dart
@@ -84,8 +84,6 @@ class SeedDao extends DatabaseAccessor<AppDatabase> with _$SeedDaoMixin {
     String commitSha = 'main',
     String commitMessage = 'feat: Test build job created by seed',
     String branch = 'main',
-    String? customScript,
-    String? workflowYaml,
   }) async {
     final now = DateTime.now().toUtc();
     final uniqueSuffix =
@@ -105,8 +103,6 @@ class SeedDao extends DatabaseAccessor<AppDatabase> with _$SeedDaoMixin {
       commitMessage: commitMessage,
       branch: branch,
       runsOn: runsOn,
-      customScript: customScript,
-      workflowYaml: workflowYaml,
       createdAt: now,
       updatedAt: now,
     );
