+++ docker-compose.yml
@@ -53,12 +53,9 @@ services:
     build:
       context: .
       dockerfile: apps/openci_build_job_processor/Dockerfile
-    container_name: openci-job-processor
     restart: unless-stopped
     environment:
       OPENCI_SERVER_URL: http://server:8080
-      OPENCI_MAX_CONCURRENT_JOBS: ${OPENCI_MAX_CONCURRENT_JOBS:-3}
-
       BASE_VM_NAME: ${BASE_VM_NAME:-tahoe-base}
 
       LOKI_URL: http://loki:3100
